+++ src/config/adwords/sources/campaign/averagefrequency.php
@@ -7,8 +7,8 @@ return [
     "fields" => [
         "AverageFrequency"
     ],
-    "parse" => function ($data): int {
-        return (int)$data->AverageFrequency;
+    "parse" => function ($data): float {
+        return (float)$data->AverageFrequency;
     },
     "inferred_from" => [
         "impressions",
