+++ src/config/adwords/sources/campaign/clickassistedconversionvalue.php
@@ -7,8 +7,8 @@ return [
     "fields" => [
         "ClickAssistedConversionValue"
     ],
-    "parse" => function ($data): int {
-        return (int)$data->ClickAssistedConversionValue;
+    "parse" => function ($data): float {
+        return (float)$data->ClickAssistedConversionValue;
     },
     "sum" => function (array $rows): float {
         return array_reduce(
